requirements.txt
langchain>=0.2.14,<0.3
langgraph>=0.2.34,<0.3
mcp>=1.2.0,<2
pydantic>=2.8,<3
pydantic-settings>=2.4,<3
typer>=0.12,<0.13
uvicorn>=0.30,<0.31
fastapi>=0.114,<0.115
httpx>=0.27,<0.28
python-dotenv>=1.0,<2
tenacity>=8.2,<9
structlog>=24.1,<25
boto3>=1.34,<2
cryptography>=43,<44

pytest>=8,<9
pytest-asyncio>=0.23,<0.24
ruff>=0.6,<0.7
black>=24.8,<25.0
